@@ -4,7 +4,7 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.12.0-0 <3.0.0'
+  sdk: '>=2.17.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -14,6 +14,7 @@ dependencies:
   flutter_triple: ^1.2.4+3
   hasura_connect:
     path: ../
+  dartz: ^0.10.1
   
 
 dev_dependencies:
