@@ -1,14 +1,15 @@
 class Task {
   final int id;
   final String title;
-  final String description;
-  final DateTime? createdAt;
-  final int idStatus;
+  // final String description;
+  // final DateTime? createdAt;
+  // final int idStatus;
 
-  Task(
-      {required this.id,
-      required this.title,
-      required this.description,
-      this.createdAt,
-      required this.idStatus});
+  Task({
+    required this.id,
+    required this.title,
+    // required this.description,
+    // this.createdAt,
+    // required this.idStatus,
+  });
 }
