@@ -1,12 +1,13 @@
-import 'package:example/app/modules/home/home_module.dart';
 import 'package:flutter_modular/flutter_modular.dart';
 import 'package:hasura_connect/hasura_connect.dart';
+
+import 'modules/home/home_module.dart';
 
 class AppModule extends Module {
   @override
   final List<Bind> binds = [
-    Bind.lazySingleton(
-        (i) => HasuraConnect('https://bwolfdev.herokuapp.com/v1/graphql')),
+    // packages binds
+    Bind((i) => HasuraConnect('https://bwolfdev.herokuapp.com/v1/graphql')),
   ];
 
   @override
