@@ -3,5 +3,9 @@ abstract class Failure implements Exception {
   final StackTrace? stackTrace;
   final Object? error;
 
-  Failure({required this.message, this.stackTrace, this.error});
+  Failure({
+    required this.message,
+    this.stackTrace,
+    this.error,
+  });
 }
