@@ -1,8 +1,6 @@
 import 'package:example/app/modules/todo/domain/entities/task.dart';
 
-class TaskMapper {
-  TaskMapper._();
-
+abstract class TaskMapper {
   static List<Task> fromJson(List<dynamic> tasks) {
     return tasks.map((e) => _make(e)).toList();
   }
@@ -10,10 +8,10 @@ class TaskMapper {
   static Task _make(Map<String, dynamic> task) {
     return Task(
       id: task['id'],
-      description: task['description'],
-      title: task['title'],
-      idStatus: task['id_status'],
-      createdAt: DateTime.tryParse(task['created_at']),
+      //description: task['description'],
+      title: task['todo'],
+      //idStatus: task['id_status'],
+      // createdAt: DateTime.tryParse(task['created_at']),
     );
   }
 }
