@@ -1,5 +1,5 @@
 import 'package:example/app/modules/todo/domain/entities/task.dart';
 
 abstract class ITaskDatasource {
-  Stream<List<Task>> watch();
+  Future<Stream<List<Task>>> watch();
 }
