@@ -5,5 +5,10 @@ import 'app/app_module.dart';
 import 'app/app_widget.dart';
 
 void main() {
-  runApp(ModularApp(module: AppModule(), child: AppWidget()));
+  runApp(
+    ModularApp(
+      module: AppModule(),
+      child: AppWidget(),
+    ),
+  );
 }
